Extract duplicated command helpers into shared functions
Moves buildFilterRules, validatePaths, checkGitRequirement, checkGitRepo,
filterBinaryFiles, resolveFillPalette, and resolveBorderMetricAndPalette
into viz_cmd_helpers.go as package-level functions. Also extracts
findGitMetric (previously in treemap_cmd.go) and introduces verifyGitRepo
as a shared core for both checkGitRequirement and checkGitRepo.

Removes ~350 lines of duplicated code across treemap, radial, bubbletree,
and spiral command files.

Fixes #192
